requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.0
dnspython==2.1.0
ecdsa==0.17.0
eventlet==0.33.0
Flask==2.0.2
Flask-AWSCognito==1.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-SocketIO==4.3.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.3.0
python-engineio==3.13.2
python-jose==3.3.0
python-socketio==4.6.0
requests==2.26.0
rsa==4.8
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
